simadb: More consistent use of albumartist info in fetch_albums_history
fetch_albums_history() returns a list of Album objects with an Artist
attribute set to albumartist (if provided, else falls back to artist).
We want an Album view of history.

fetch_albums_history(needle) returns a list of Album objects with
Artist attribute set to artist (in this case we want album played for
artist == needle). Here we want the artist view on album history.